requirements.txt
dj_config_url==0.1.1
Django==5.1.3
djangorestframework==3.15.2
environ==1.0
